+++ packages/core/src/telemetry.ts
@@ -37,7 +37,6 @@ import {
   ATTR_SERVICE_NAME,
   ATTR_SERVICE_VERSION,
 } from '@opentelemetry/semantic-conventions';
-import 'zone.js';
 import { WayfinderEmitter } from './emitter.js';
 import type {
   GatewaysProvider,
@@ -45,17 +44,25 @@ import type {
   WayfinderOptions,
 } from './types.js';
 import { isBrowser, isChromeExtension } from './utils/browser.js';
+import { assertZoneLoaded, loadZonePolyfill } from './utils/zone.js';
 import { WAYFINDER_CORE_VERSION } from './version.js';
 
 // avoid re-initializing the tracer provider and tracer
 let tracerProvider: WebTracerProvider | NodeTracerProvider | undefined;
 let tracer: Tracer | undefined;
 
+// load zone.js polyfill if it's not already loaded
+if (isBrowser()) {
+  loadZonePolyfill();
+}
+
 export const initTelemetry = ({
   enabled = false,
   sampleRate = 0.1, // 10% sample rate by default
   exporterUrl = 'https://api.honeycomb.io/v1/traces',
   apiKey = 'c8gU8dHlu6V7e5k2Gn9LaG', // intentionally left here - if it gets abused we'll disable it
+  clientName,
+  clientVersion,
 }: TelemetrySettings):
   | {
       tracerProvider: WebTracerProvider | NodeTracerProvider;
@@ -86,6 +93,8 @@ export const initTelemetry = ({
   const resource = resourceFromAttributes({
     [ATTR_SERVICE_NAME]: 'wayfinder-core',
     [ATTR_SERVICE_VERSION]: WAYFINDER_CORE_VERSION,
+    'client.name': clientName,
+    'client.version': clientVersion,
   });
 
   const useWebTracer = isBrowser() || isChromeExtension();
@@ -105,6 +114,11 @@ export const initTelemetry = ({
         resource,
         spanProcessors: [spanProcessor],
       });
+
+  // ensure zone.js is loaded before registering the tracer provider if we're using the browser
+  if (useWebTracer) {
+    assertZoneLoaded();
+  }
 
   provider.register({
     // zone.js is only used in the browser (node/extensions/service workers don't need it)